requirements.txt
# Django
django===4.1.1
# Django Server
daphne===3.0.2
asgiref===3.5.2
# Kit utilities
django-extensions===3.1.3
# Django Channels
channels==3.0.5
# Redis Layer
channels_redis===3.4.1
# Requests
requests===2.28.1
# Observability
opentelemetry-api==1.13.0
opentelemetry-sdk==1.13.0
opentelemetry-instrumentation-asgi==0.34b0
## Exporters
opentelemetry-exporter-jaeger==1.13.0
django-prometheus==2.2.0
